+++ src/main/java/org/sopt/domain/auth/infrastructure/persistence/RefreshTokenRepositoryImpl.java
@@ -1,4 +1,4 @@
-package org.sopt.domain.auth.infrastructure;
+package org.sopt.domain.auth.infrastructure.persistence;
 
 import lombok.RequiredArgsConstructor;
 import org.sopt.domain.auth.domain.model.RefreshToken;
