+++ src/main/java/org/sopt/domain/auth/infrastructure/persistence/SocialAccountJpaRepository.java
@@ -1,4 +1,4 @@
-package org.sopt.domain.auth.infrastructure;
+package org.sopt.domain.auth.infrastructure.persistence;
 
 import org.sopt.domain.auth.domain.model.OAuthProvider;
 import org.sopt.domain.auth.domain.model.SocialAccount;
